Rename devnet.rs to background_devnet.rs
