+++ crates/starknet-server/tests/common/mod.rs
@@ -1,3 +1,5 @@
+pub mod background_devnet;
 pub mod constants;
+pub mod data;
 pub mod devnet;
 pub mod utils;
