+++ crates/starknet-server/tests/test_get_class.rs
@@ -6,8 +6,8 @@ mod get_class_tests {
         MaybeUnknownErrorCode, Provider, ProviderError, StarknetErrorWithMessage,
     };
 
+    use crate::common::background_devnet::BackgroundDevnet;
     use crate::common::constants::PREDEPLOYED_ACCOUNT_ADDRESS;
-    use crate::common::devnet::BackgroundDevnet;
 
     #[tokio::test]
     async fn test_get_class_at() {
