+++ crates/starknet/src/starknet/starknet_config.rs
@@ -50,6 +50,7 @@ impl Default for StarknetConfig {
             predeployed_accounts_initial_balance: DEVNET_DEFAULT_INITIAL_BALANCE.into(),
             host: DEVNET_DEFAULT_HOST,
             port: DEVNET_DEFAULT_PORT,
+            start_time: None,
             timeout: DEVNET_DEFAULT_TIMEOUT,
             gas_price: DEVNET_DEFAULT_GAS_PRICE,
             chain_id: DEVNET_DEFAULT_CHAIN_ID,
